# Conflicts:
#	api-impl/src/main/resources/templates/index.ftl
#	common/src/main/java/com/toutiao/web/common/util/ESClientTools.java

+++ dao/src/main/java/com/toutiao/web/dao/mapper/officeweb/TotalListedRatioMapper.java
@@ -15,5 +15,5 @@ public interface TotalListedRatioMapper extends BaseDao {
 
     int insertSelective(TotalListedRatio record);
 
-    Double selectByTotalPrice(@Param("plotTotalFirst")Integer plotTotalFirst,@Param("plotTotalEnd")Integer plotTotalEnd);
+    Double selectByTotalPrice(@Param("plotTotalFirst")Double plotTotalFirst,@Param("plotTotalEnd")Double plotTotalEnd);
 }
